Refactoring for best practices and code organization. Added user management features, including user roles and permissions. Updated security configuration to support role-based access control. Improved error handling and added new endpoints for user management in the API. Updated frontend to include user management UI components and actions.

+++ api/src/main/kotlin/com/joaohonorato/blog/model/GithubRepo.kt
@@ -13,7 +13,7 @@ class GithubRepo(
     val id: Int? = null,
 
     @Column(name = "github_id", unique = true, nullable = false)
-    val githubId: Int = 0,
+    val githubId: Long = 0L,
 
     @Column(nullable = false)
     val name: String = "",
